move healthchecks to correct modules
