+++ core/healthcheck/module.go
@@ -6,7 +6,6 @@ package healthcheck
 
 import (
 	"flamingo.me/dingo"
-	"flamingo.me/flamingo/v3/core/healthcheck/domain/healthcheck"
 	"flamingo.me/flamingo/v3/core/healthcheck/interfaces/controllers"
 	"flamingo.me/flamingo/v3/framework/prefixrouter"
 	"flamingo.me/flamingo/v3/framework/systemendpoint"
@@ -16,31 +15,22 @@ import (
 
 // Module entry point for the flamingo healthcheck module
 type Module struct {
-	controller      *controllers.Healthcheck
-	checkSession    bool
-	checkAuthServer bool
-	checkPath       string
-	pingPath        string
-	sessionBackend  string
+	controller *controllers.Healthcheck
+	checkPath  string
+	pingPath   string
 }
 
 // Inject dependencies
 func (m *Module) Inject(
 	controller *controllers.Healthcheck,
 	config *struct {
-		CheckSession    bool   `inject:"config:core.healthcheck.checkSession"`
-		CheckAuthServer bool   `inject:"config:core.healthcheck.checkAuth"`
-		CheckPath       string `inject:"config:core.healthcheck.checkPath"`
-		PingPath        string `inject:"config:core.healthcheck.pingPath"`
-		SessionBackend  string `inject:"config:flamingo.session.backend"`
+		CheckPath string `inject:"config:core.healthcheck.checkPath"`
+		PingPath  string `inject:"config:core.healthcheck.pingPath"`
 	},
 ) {
 	m.controller = controller
-	m.checkSession = config.CheckSession
-	m.checkAuthServer = config.CheckAuthServer
 	m.checkPath = config.CheckPath
 	m.pingPath = config.PingPath
-	m.sessionBackend = config.SessionBackend
 }
 
 type routes struct {
@@ -53,25 +43,11 @@ func (r *routes) Inject(controller *controllers.Ping) {
 
 func (r *routes) Routes(registry *web.RouterRegistry) {
 	registry.HandleAny("core.healthcheck.ping", web.WrapHTTPHandler(r.controller))
-	registry.Route("/health/ping", "core.healthcheck.ping")
+	registry.MustRoute("/health/ping", "core.healthcheck.ping")
 }
 
 // Configure dependency injection
 func (m *Module) Configure(injector *dingo.Injector) {
-	if m.checkSession {
-		switch m.sessionBackend {
-		case "redis":
-			injector.BindMap(new(healthcheck.Status), "session").To(healthcheck.RedisSession{})
-		case "file":
-			injector.BindMap(new(healthcheck.Status), "session").To(healthcheck.FileSession{})
-		default:
-			injector.BindMap(new(healthcheck.Status), "session").To(healthcheck.Nil{})
-		}
-	}
-	if m.checkAuthServer {
-		injector.BindMap((*healthcheck.Status)(nil), "auth").To(healthcheck.Auth{})
-	}
-
 	injector.BindMap((*domain.Handler)(nil), m.pingPath).To(&controllers.Ping{})
 	injector.BindMap((*domain.Handler)(nil), m.checkPath).To(&controllers.Healthcheck{})
